requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
frozenlist==1.3.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
multidict==5.2.0
mypy==0.931
mypy-extensions==0.4.3
oauthlib==3.1.1
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tomli==2.0.0
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
yarl==1.7.2
